refactor: move mail files into service module
