+++ src/app/modules/bounce/__tests__/bounce-test-helpers.ts
@@ -1,7 +1,7 @@
 import { ObjectId } from 'bson'
-import { cloneDeep, merge, pick } from 'lodash'
+import { merge, pick } from 'lodash'
 
-import { EmailType } from 'src/app/constants/mail'
+import { EmailType } from 'src/app/services/mail/mail.constants'
 import {
   BounceType,
   IBounce,
