+++ src/app/modules/bounce/__tests__/bounce.controller.spec.ts
@@ -5,12 +5,12 @@ import dbHandler from 'tests/unit/backend/helpers/jest-db'
 import expressHandler from 'tests/unit/backend/helpers/jest-express'
 import { mocked } from 'ts-jest/utils'
 
-import { EmailType } from 'src/app/constants/mail'
 // eslint-disable-next-line import/first
 import { handleSns } from 'src/app/modules/bounce/bounce.controller'
 import getBounceModel from 'src/app/modules/bounce/bounce.model'
 import * as BounceService from 'src/app/modules/bounce/bounce.service'
 import * as FormService from 'src/app/modules/form/form.service'
+import { EmailType } from 'src/app/services/mail/mail.constants'
 import { IBounceSchema, ISnsNotification } from 'src/types'
 
 const Bounce = getBounceModel(mongoose)
