+++ src/app/modules/bounce/__tests__/bounce.service.spec.ts
@@ -8,16 +8,11 @@ import dbHandler from 'tests/unit/backend/helpers/jest-db'
 import getMockLogger from 'tests/unit/backend/helpers/jest-logger'
 import { mocked } from 'ts-jest/utils'
 
-import { EMAIL_HEADERS, EmailType } from 'src/app/constants/mail'
 import getFormModel from 'src/app/models/form.server.model'
-import MailService from 'src/app/services/mail.service'
+import { EMAIL_HEADERS, EmailType } from 'src/app/services/mail/mail.constants'
+import MailService from 'src/app/services/mail/mail.service'
 import * as LoggerModule from 'src/config/logger'
-import {
-  BounceType,
-  IFormSchema,
-  ISnsNotification,
-  IUserSchema,
-} from 'src/types'
+import { BounceType, ISnsNotification, IUserSchema } from 'src/types'
 
 import { makeBounceNotification, MOCK_SNS_BODY } from './bounce-test-helpers'
