+++ src/app/services/mail/__tests__/mail.service.spec.ts
@@ -3,17 +3,15 @@ import moment from 'moment-timezone'
 import Mail, { Attachment } from 'nodemailer/lib/mailer'
 import { ImportMock } from 'ts-mock-imports'
 
-import { MailSendError } from 'src/app/modules/mail/mail.errors'
-import { MailService } from 'src/app/services/mail.service'
-import * as MailUtils from 'src/app/utils/mail'
+import { MailSendError } from 'src/app/services/mail/mail.errors'
+import { MailService } from 'src/app/services/mail/mail.service'
 import {
   AutoreplySummaryRenderData,
-  BounceType,
-  IPopulatedForm,
-  ISubmissionSchema,
   MailOptions,
   SendAutoReplyEmailsArgs,
-} from 'src/types'
+} from 'src/app/services/mail/mail.types'
+import * as MailUtils from 'src/app/services/mail/mail.utils'
+import { BounceType, IPopulatedForm, ISubmissionSchema } from 'src/types'
 
 const MOCK_VALID_EMAIL = 'to@example.com'
 const MOCK_VALID_EMAIL_2 = 'to2@example.com'
