+++ src/app/services/mail/mail.constants.ts
