+++ tests/unit/backend/modules/verification/verification.service.spec.ts
@@ -12,7 +12,7 @@ import {
   resetFieldInTransaction,
   verifyOtp,
 } from 'src/app/modules/verification/verification.service'
-import MailService from 'src/app/services/mail.service'
+import MailService from 'src/app/services/mail/mail.service'
 import { SmsFactory } from 'src/app/services/sms/sms.factory'
 import { generateOtp } from 'src/app/utils/otp'
 import formsgSdk from 'src/config/formsg-sdk'
